Move configuration logic from MasterNodeApp, add comments
